+++ app/controllers/patients/pipelines_controller.rb
@@ -64,7 +64,7 @@ module Patients
     end
 
     def load_patient
-      @patient = Patient.where(case_number: params[:patient_id]).take
+      @patient = Patient.find_by!(case_number: params[:patient_id])
     end
 
     def find_and_authorize